requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1; python_version < '3.9'
django-filter==23.3
django==4.1.3
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
importlib-metadata==6.8.0; python_version < '3.10'
inflection==0.5.1
markdown==3.4.4
packaging==23.1
psycopg2-binary==2.9.7
pyjwt==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
sqlparse==0.4.4
typing-extensions==4.8.0; python_version < '3.11'
uritemplate==4.1.1
zipp==3.17.0
